Update patch set 8
Patch Set 8:

Patch set 8 tweaks the horizontal scroll limits. It uses the extraSpace from hrev24130 ('till now) and reverses it and also makes it also apply to B_LEFT_ALIGNED text view's as well. This change makes it so that half of the text is visible in the text view box after you have added enough text to scroll the text rect off the end of the available space. I'm pretty sure that was the intension of hrev24130 but things have changed since then. So this restores the intended behavior of hrev24130. I was using a different value (font size * 3), but I think the half text view width that stippi chose is better, so I have restored it.

Patch-set: 8
Reviewer: Gerrit User 1000013 <1000013@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>