Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> > You can load them using the DefaultCatalog class (in headers/private/locale) but adding something to the LocaleRoster and/or BCatalog to do it more directly is probably a good idea (a constructor taking a MIME type instead of an entry ref would do).
> > 
> > What will be lost is the ability to load a catalog that would be embedded into the executable, if we do it this way.
> 
> Why? If we add a new constructor the old way would still work, wouldn't it? Am I misunderstanding something?

The existing constructor will work as before, yes. What I meant is the new constructor we'll be adding will only look in /system/data/locale/catalogs, and will not be able to load a catalog that would be stored in some application resource. This makes things possibly confusing: if you look at the function prototypes, with one taking a MIME type and the other taking an entry_ref, it could be expected that they do the same thing (the entry_ref being just analyzed to find the corresponding MIME type). Well, I'll see if I cna do something about it, or otherwise I'll just document it.

> 
> > Do you want to make the changes to locale kit, or should I look into it?
> 
> Since you know the code better it would be more efficient if you prepared the changes :)

Ok, I'll have a look at it.

Patch-set: 1