Update patch set 3
Patch Set 3:

> Patch Set 3:
> 
> I think I found something that may help here: The 'AutoTextControl' from libWalter in the HaikuArchives. It's MIT licensed and handles limiting characters through a BMessageFilter.
> 
> Quoting a portion of the description: "...AutoTextControl can set a limit on the number of characters in the text box, which is probably more often what you really want to do instead of limiting by byte count."
> 
> You can find the code here: https://github.com/HaikuArchives/LibWalter/blob/develop/src/general/AutoTextControl.cpp

Now thinking about it, this functionality should probably be integrated directly into the BTextView class. After all, this wouldn't be too difficult to add, and it wouldn't break compatibility with BeOS apps or existing Haiku apps. However, we may want to avoid the BMessageFilter part, and handle the check the same way the "disallowed characters" are checked.

Patch-set: 3