Update patch set 10
Patch Set 10:

> At the point when Tracker asks for a thumbnail, I don’t have any information about what size icon has been stored in an attribute, I only know what size icon I’m being asked to generate. So I’d have to store that information, and then retrieve it in order to translate the PNG to a BBitmap at the size it’s at, then convert back to a PNG at the size we want. I could store the attribute with :64 on the end so that I know it’s a 64x64 icon, or I could store “which” (the icon size) in a different attribute.

The idea with a single size stored would be:
- Get the png from the attribute, you get a BBitmap at the size the PNG is in
- Draw it using DrawBitmap which does both the drawing and the resizing
- If you notice the size doesn't match the current setting, regenerate the thumbnail (maybe only if the new size is larger)

So, it isn't that hard either way.

Advantages for storing only a single size:
- Saves disk space
- More future proof in the sense that we can add new sizes without problems (or remove some, even)
- Faster initial generation since we only need ot generate one size

Advantages for storing multiple sizes:
- Possibly nicer looking thumbnails when changing the view size
- Generation of thumbnails happens only once, not everytime you zoom in and out
- Avoids scaling down the picture to generate the thumbnail, and then again to render it

I'd say multiple size are probably useful. I imagine I would glance at a set of photos with small-ish icons, but then zoom in to see them a bit closer once I located roughly the things I'm after and I look for a specific picture or detail.

Basically, currently we all use fixed size icons, but the fact that there are thumbnails means we will probably want to zoom in and out a lot more and more often. So having the multiple sizes makes sense. And I guess they wouldn't even be all that big anyway, even if storing multiple ones.

Ideally we would use an image format that allows storing multiple resolutions in a single file and avoids some redundancy.

> I can store a thumbnail creation time that is more accurate by using real_time_clock_usecs() but I am still comparing that value to a 32-bit time_t in modification time, so this doesn’t solve the 2038 timebomb, unless I’m missing something.

You can use real_time_clock_usecs also for the comparison and completely remove time_t. Probably the safest thing to do indeed, we get both increased resolution and longer timespan.

> One last thing, since I’m storing the bitmaps as raw PNG data, I may have to create some more Tracker constants for endianness, and then convert the endianness to the one on your system when reading the data. This is will be necessary to share thumbnails between little endian and some future (or past, PPC) big endian Haiku.

I don't understand, PNG is the same no matter which platform you read it on? (the P is for "Portable", after all). PNG can encode several colorspaces and libpng (or the png translator, in our case) handle the details for you, so you don't need to worry about that?

Patch-set: 10