Update patch set 10
Patch Set 10:

> Patch Set 10:
> > I can store a thumbnail creation time that is more accurate by using real_time_clock_usecs() but I am still comparing that value to a 32-bit time_t in modification time, so this doesn’t solve the 2038 timebomb, unless I’m missing something.
> 
> You can use real_time_clock_usecs also for the comparison and completely remove time_t. Probably the safest thing to do indeed, we get both increased resolution and longer timespan.

Can you elaborate on this because I still don't understand. I am comparing the thumbnail creation time of the thumbnail I generated to the modification time of the file. The thumbnail creation time is not the problem, it is always 8-bytes, but the file's modification time is 4-bytes on 32-bit Haiku. How can I use real_time_clock_usecs() for the comparison and remove the time_t? The modification time is stored in an attribute as a 4-byte time_t and I have no control over this. If I were to convert the modification time to usecs using real_time_clock_usecs() the value would have overflowed after 2038 and give me the wrong value. 

> > One last thing, since I’m storing the bitmaps as raw PNG data, I may have to create some more Tracker constants for endianness, and then convert the endianness to the one on your system when reading the data. This is will be necessary to share thumbnails between little endian and some future (or past, PPC) big endian Haiku.
> 
> I don't understand, PNG is the same no matter which platform you read it on? (the P is for "Portable", after all). PNG can encode several colorspaces and libpng (or the png translator, in our case) handle the details for you, so you don't need to worry about that?

PNGs are portable yes, but AFAIK this endian issue doesn't have anything to do with PNG. I'm storing the thumbnail bytes into an attribute, and those bytes are stored in little endian order. When they are read back from the attribute, they are read in machine order, and if machine endianness doesn't match the endianness of the stored attribute, the resulting PNG image will be wrong. Unless there is some factor that I don't comprehend that will make this work.

> Patch Set 10:
> Another question, do we plan to have thumbnails for non-image files (for example, videos?) do we need translators handling video files to do that? What about other filetypes (PDF? LibreOffice files?)

Yeah maybe! As long as we have a data translator for the file we can use it to generate a thumbnail. For videos, we could extract a screenshot, but the timestamp would have to be chosen carefully not to show a black image (as the first frame of videos is often black.) Do we have a data translator for PDF? If so this thumbnail generation might already work for PDFs. We can put generating thumbnails for these file types in the "future enhancements" bucket.

Patch-set: 10
Reviewer: Gerrit User 1000013 <1000013@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>