Update patch set 10
Patch Set 10:

> Patch Set 10:
> PNGs are portable yes, but AFAIK this endian issue doesn't have anything to do with PNG. I'm storing the thumbnail bytes into an attribute, and those bytes are stored in little endian order. When they are read back from the attribute, they are read in machine order, and if machine endianness doesn't match the endianness of the stored attribute, the resulting PNG image will be wrong. Unless there is some factor that I don't comprehend that will make this work.

The PNG format explicitly defines what endianness things are to be stored in where it matters. The bytes on disk will be 100% identical regardless of whether they were written by a big or little endian machine, or else it isn't a valid PNG. This is the very definition of a portable format.

Patch-set: 10
CC: Gerrit User 1000015 <1000015@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>