Update patch set 10
Patch Set 10:

> Can you elaborate on this because I still don't understand. I am comparing the thumbnail creation time of the thumbnail I generated to the modification time of the file. The thumbnail creation time is not the problem, it is always 8-bytes, but the file's modification time is 4-bytes on 32-bit Haiku. How can I use real_time_clock_usecs() for the comparison and remove the time_t? The modification time is stored in an attribute as a 4-byte time_t and I have no control over this. If I were to convert the modification time to usecs using real_time_clock_usecs() the value would have overflowed after 2038 and give me the wrong value. 

Ah, yes, of course, I was thinking of the current time and not the file modification time.

Interestingly, BFS already uses 64bit to store the modification time, so we could add APIs (in BStatable and probably layers under it) to get the full value.

> 
> > > One last thing, since I’m storing the bitmaps as raw PNG data, I may have to create some more Tracker constants for endianness, and then convert the endianness to the one on your system when reading the data. This is will be necessary to share thumbnails between little endian and some future (or past, PPC) big endian Haiku.
> > 
> > I don't understand, PNG is the same no matter which platform you read it on? (the P is for "Portable", after all). PNG can encode several colorspaces and libpng (or the png translator, in our case) handle the details for you, so you don't need to worry about that?
> 
> PNGs are portable yes, but AFAIK this endian issue doesn't have anything to do with PNG. I'm storing the thumbnail bytes into an attribute, and those bytes are stored in little endian order. When they are read back from the attribute, they are read in machine order, and if machine endianness doesn't match the endianness of the stored attribute, the resulting PNG image will be wrong. Unless there is some factor that I don't comprehend that will make this work.

If you are storing bytes, they are stored in the order you put them. ASCII and UTF8 strings do not have endianness problems, for example. You get problems only if you manipulate int32 or int16 values. And even then, bfs will take care of swapping them for you (that's a reason for attributes to be typed).

You store your PNG in the attribute, and it's read as-is. Just like reading a PNG file from disk wouldn't need anything specific.

Patch-set: 10