Update patch set 10
Patch Set 10:

> Patch Set 10:
> Advantages for storing multiple sizes:
> - Possibly nicer looking thumbnails when changing the view size
> - Generation of thumbnails happens only once, not everytime you zoom in and out
> - Avoids scaling down the picture to generate the thumbnail, and then again to render it
> 
> I'd say multiple size are probably useful. I imagine I would glance at a set of photos with small-ish icons, but then zoom in to see them a bit closer once I located roughly the things I'm after and I look for a specific picture or detail.

I doubt you'll see a difference between an original 16x16 or even 32x32 or a downscaled one. And even if, the difference won't be worth it.

You missed a huge disadvantage of this: space is not just precious because it's nice to be lean. BFS stores attributes in the inode, as long as space allows. If the thumbnail is stored more than once, chances are high that this space won't suffice, which means the attribute will be outsourced to a file. This can be a huge speed impact when opening a folder, as inodes within a folder will be close to each other on disk. There is no space for the attributes, though, they will reside at some other place. This is especially true with HDs, but will also be noticeable with SSDs. This is also why HVIF has been invented.
Of course, it would be helpful to know the usual size of PNGs to make a more educated decision -- if they are always too large, this wouldn't make much of a difference.

I would just use a minimal size, like 64 pixels for the long edge, and only store larger sizes when you actually increase the icon size. The generation of thumbnails would also not happen every time you zoom in and out, but only once, like with the other solution, or even less often, as there is a minimum thumbnail size.

> You can use real_time_clock_usecs also for the comparison and completely remove time_t. Probably the safest thing to do indeed, we get both increased resolution and longer timespan.

I haven't looked, but I'm pretty sure there is a way to get more than time_t precision out of the file system API.

Patch-set: 10
Reviewer: Gerrit User 1000016 <1000016@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>