Update patch set 10
Patch Set 10:

> Patch Set 10:
> 
> > Patch Set 10:
> > 
> > > Patch Set 10
> > > 
> > > You missed a huge disadvantage of this: space is not just precious because it's nice to be lean. BFS stores attributes in the inode ...
> > 
> > Re-reading what you said, I take it back, you do get the benefits described. My complaint is mostly about bitmap scaling artifacts, so you are right that storing a smaller attribute will be faster ...
> 
> Here are the results of creating a 64x64 thumb directly vs creating a 32x32 thumb and then scaling it back 64x64, YUCK!
> 
> https://64.media.tumblr.com/2d3adc3ace32063459dfc6819149f708/9e1a6899df24fba0-a1/s640x960/0435aedb0aacad382bf2d26dab169e1fbab351c2.png

But that would never be visible to the user. If our minimal stored icon size is 64, then you would only get blurry images when you enlarge your view to 128 pixels (or even more in the future). But that would trigger redoing that thumbnail in the larger size, and then it would become sharp while looking at it.

Ie. the logic would be similar to this:
1) Does the file has Media:Thumbnail?
1.yes) Is it current?
    2.yes) then show it at the requested size!
        3) Is the image too small?
            3.yes) go to 4)
1.no) go to 4)
4) Create a new thumbnail asynchronously at std::max(requestedSize, minimalSize), and present it to the user.

Patch-set: 10
Reviewer: Gerrit User 1000016 <1000016@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>