Update patch set 10
Patch Set 10:

> Patch Set 10:
> 
> The real difference is that my way saves multiple attributes for different sizes and your want to cleverly reuses the same attribute, but the benefit of doing so is not worth it. If you switch a folder full of images between 16x16 list view and 128x128 icon view and back again my way does a lot less work than your way. And what’s the benefit of doing it your way? I still need to generate 128x128 thumbnails even though they don’t fit in the inode. I can’t upscale or it looks like garbage. If I downscale then I have to store a bigger size attribute and shrink it. I’d rather store 1 thumb per size.

I think you still don't fully understand all the implications. My suggestion actually needs to do a lot less; if I switch from 16 to 128 pixels, my solution needs to compute two thumbnails (one in 64, one in 128 bytes). Your solution does the same (one in 16, one in 128 bytes).

However, if I switch from 16 to 32 to 64 to 128, your solution needs to compute 4 thumbnails, while mine stays as 2. If we go the other way around, my solution would only ever need to compute 1 thumbnail, while yours create up to 4.

And in addition to all that, the chances that one thumbnail (no matter the size) will fit into the inode is a lot higher than that 4 thumbnails will fit there.

Anyway, I just checked (only with a single photo, so it's not very representative), how large PNGs get (I used PNG-8 which was considerably smaller than PNG-24):

width size
128   7 KB
64    3.5 KB
32    1.5 KB
16    750 B

With an inode size of 2 KB or 4 KB (the latter is what should be used with todays HDs, and SSDs), we have between 1.5, and 3.5 KB space for attributes. Say we need 500 of that for other attributes, like stuff your favourite photo database stores there. That leaves us with 1 KB or 3 KB of space. That would only ensure space for either 16 pixel or 32 pixel large images. I think that even 128 pixel is very small for actually looking that thumbnails, so if that will be the most common choice, you can pretty much forget the debate about storing the attribute in the inode.

However, in list mode, a 16x16 image could be put in there. But even that isn't necessarily the size it'll be, as with larger font sizes, we also use larger images there. In any case, it has a good chance to find its space in the attributes.

So using that as a base, it would actually make sense to store two different sizes to be able to make use of that attribute space. I would consider anything else as wasteful without any advantages, though -- having to create 4 different thumbnails instead of just two does not give you a single advantage in reality, as you can always use the same scaling mechanism (no matter if you go from 4000 pixels to 32 or from 128 to 32, and the results will be very similar).

Patch-set: 10
Reviewer: Gerrit User 1000016 <1000016@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>