Update patch set 4
Patch Set 4:

I did the following test today:
- Build a clean Haiku image and boot it
- Start the Input preferences

With this change, the Input preferences gets corrupt settings for the mouse. First I had a button mapped to "P" instead of a number, and the preferences crashed when I tried to change it. Then I tried again and I had a mouse with 0 buttons, and another with apparently a very large number.

This is not fixed by the next commits, and with the last one I still get a similar error. This should be fixed before going further.

Patch-set: 4
Reviewer: Gerrit User 1000001 <1000001@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>