Update patch set 1
Patch Set 1:

I would at least split this into two controls, one to select the position, one to select the mode (mini or not). Wether it's a menu or a set of radio buttons, having a setting with 14 possible states will be confusing. You can try it but I doubt 14 unlabelled radio buttons is a large improvement over 14 menu items with imprecise names.

It would look like this, I think:

 +--(r)--(r)--(r)--+
 |                 |
(r)               (r)
 |                 |
(r)               (r)
 |                 |
 |                 |
(r)               (r)
 |                 |
(r)               (r)
 |                 |
 +--(r)--(r)--(r)--+

We need to group the settings to make sense of them:

Normal mode:
   - Top
   - Bottom
   - Top left
   - Top right
   - Bottom left
   - Bottom right

Vertical mini mode:
   - Left top
   - Left bottom
   - Right bottom
   - Right top

Horizontal mini mode:
   - Top left
   - Top right
   - Bottom left
   - Bottom right

So now we can have a "mode" menu (with just 3 options), and checkboxes (4 or 6 of them depending on the mode) or maybe some kind of "compass" control that you could put in several possible orientations. With 6 orientations it's still a complex control, but it's a bit more reasonable already and you don't need a 128x128 px area for a single control then.

Patch-set: 1
CC: Gerrit User 1000001 <1000001@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>