Update patch set 13
Patch Set 13:

> This is what the BeBook calls a “message cache” [0] which is a single entry message queue. If calling receive_data(&sender, NULL, 0) clears the message cache, and I think you are right that it does, then the code I wrote will not work as I’m eating the icon data when I check to make sure it’s mine. I only recently learned about these methods, and reading further a port seems to be what I need here which is a multi-level cache.

Is there a reason for going with the low-level APIs here? I would just start a BLooper and use BMessage to send the data back and forth. Or possibly look into BJob and related classes which were designed for this pattern during the package kit work, but are currently undocumented.

Patch-set: 13