Update patch set 13
Patch Set 13:

> Patch Set 12:
> 
> As much fun as attributes are, I still need to fix up the non-attribute code path. I burned an ISO9660 CD of images on my Mac that knows nothing about BeOS mime types. When I load the CD, all the icons show up as folder or generic file. I believe this is because the volume does not support mime. All of the files claim their mime type as application/x-vnd.Be-directory. 

That looks wrong. We should be able to get files vs directories right, at least?

For some filesystems (I think FAT?), the filesystem itself will actually generate the mimetype attribute from the extension, to fake some minimal support for file types. I checked the iso9660 sourcecode and it does not do this (all attribute hooks are set to NULL).

But even if that's not done, the MIME sniffing code should still work, it's just that the results can't be stored on disk. So the icon should appear when Tracker identifies the file, and then revert to the default one when the window is closed, or maybe when the icon is dropped from the icon cache or something.

How did you check the MIME type? I'd expect nothing at all would show in attributes for iso9660? If there's something there, we have to understnad where it's coming from.

Patch-set: 13