Update patch set 14
Patch Set 14:

> Patch Set 11:
> 
> > Patch Set 11:
> > 
> > Ok I give in. I wrote most of the code already anyway. Patch set 10 stores a minimum icon size of 64x64, down-scales that for smaller and stores a separate attribute for 96x96 and 128x128. Sizes of my sample image 64x64 thumbnails vary from 1.2k to 8k. If we need a smaller minimum size we can do that.
> 
> Well, you either didn't really read what I wrote, or you have a different understanding of giving in :-)

I implemented the feature you wanted by doing an extra conversion, the difference that remains is the decision on which icon sizes we want to stash.

> Storing multiple images is and stays wasteful for no gain, unless it fits in the inode data, and that will be true only for the list size (ie. around 16x16 pixel). Otherwise, I would always just store the largest size Tracker supports, and scale that down. Why would you want to store a 96 pixel size besides a 64 pixel wide one? Therefore, I would also have only two attributes for this, "Media:ListThumbnail", and "Media:Thumbnail" -- not mentioning any size on purpose, because it's not set into stone anywhere.
> Looking up attributes takes time, too!

Small detail, 16x16 is also used in mini icon mode so it wouldn’t be ListThumbnail but MiniThumbnail,

Currently I’m stashing 16x16, 32x32, 64x64 and 128x128, and downscaling for the rest. Your idea is to stash 16x16 and 128x128 and downscale to get the rest.

128x128 is a lot bigger than 32x32, so if you want to keep your system nice and lean, you can stick to the BeOS sizes and hopefully the icon will fit in the inode my way, but your way once you’re in icon mode the inode is blown. So what remains then is 64x64, this could be eliminated, but I like that the currently stashed sizes are all integer multiples of each other, and losing 64x64 would mean downscaling 128x128 to 40x40.

There is a small decrease in quality from downscaling, not as great as the upscaling was from earlier, but still noticeable. If you wait long enough I can provide a screen shot comparison.

Patch-set: 14
Reviewer: Gerrit User 1000013 <1000013@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>