Update patch set 6
Patch Set 6:

Sorry for the many patch sets, this turned out to be bigger than I thought it was going to be. I just noticed that I remove a space between block_run and the ()’s in a couple of instances. But other instances omit the space so I think that’s fine. There is no consistency between when it separates the printed values with .’s and ,’s that I can tell, but whatever, if there is a logical difference, I’ve preserved it.

Patch-set: 6
Reviewer: Gerrit User 1000013 <1000013@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>