Update patch set 2
Patch Set 2:

> I am not against removing the limit altogether. I am just unsure to what extend we want to (1)  limit bad behaviour by the developers (i.e. join your threads or detach them! Don't wait forever for them) and (2) protect the system (and kernel) from intentional or accidental misuse.

What is "bad behaviour" here? If the developer decides that creating thousands of threads and not joining them quickly is ok, why would that be bad? The only reason I see is that it could make pthread_join slower as it needs to go through this list, and maybe memory increase as we need to store the list entries, but I don't think they are that large anyway.

The second point would be more interesting to look at. Could a developer, on purpose, use this to fill the kernel memory space by creating and never joining a lot of threads? If so, then yes, we definitely need to see if we can do something.

Otherwise, by the same logic we could say "we limit memory allocations to 1MB, if you use more, you probably have a memory leak in your code and that is bad". Doesn't sound great, right? And, we would handle this, not by preventing malloc to work, but by freeing random parts of the memory to make space for new allocations (just as we are dropping old non-joined threads here).

If we want to have a limit, we need to enforce it at thread creation time. Then we can return an error in pthread_create and the app has a chance to handle it. POSIX allows us, for example, to take these "dead" threads into account when enforcing PTHREAD_THREADS_MAX (source: https://pubs.opengroup.org/onlinepubs/9699919799/functions/pthread_join.html). So then pthread_create could fail because there are too many pending or running joinable threads and it's not allowed to create more. That puts us back into a normal error-handling situation where we just return an error to the app and the app can handle it as it wants.

But here, we are just making pthread_join behave strangely and forget about some threads. We now try to go around this problem by having a message in the syslog, which applications have no way to catch, but developers may see and act upon.

So, it's not that I'm against having a limit (and indeed, to protect the kernel from intentional misuse, we probably do need one somewhere), but more that the limit is enforced at the wrong place, making it impossible for applications to gracefully handle the error, and resulting into confusing behavior for app developers.

Patch-set: 2