Update patch set 16
Patch Set 16:

> Patch Set 13:
> 
> > This is what the BeBook calls a “message cache” [0] which is a single entry message queue. If calling receive_data(&sender, NULL, 0) clears the message cache, and I think you are right that it does, then the code I wrote will not work as I’m eating the icon data when I check to make sure it’s mine. I only recently learned about these methods, and reading further a port seems to be what I need here which is a multi-level cache.
> 
> Is there a reason for going with the low-level APIs here? I would just start a BLooper and use BMessage to send the data back and forth. Or possibly look into BJob and related classes which were designed for this pattern during the package kit work, but are currently undocumented.

Using a port allows me to send the icon data back to IconCache and then pick it up there when the window thread comes back though. Once I got Tracker to recognize image mime types on read-only volumes, i was able to test my port code and make it work.

Once the icon is generated, I send a BMessage containing the node_ref from the generator thread to be_app to trigger the icon to update. Because I don’t have access to the Pose in IconCache (or pose view or window), I am routing the message through the Tracker application thread, which tells the window to tell the pose view to tell the pose to update. I’d like to eliminate this if I can figure out how to and update the pose directly from the model, but that doesn’t seem to be possible, possibly intensionally. A model is a wrapper around a BNode, and a BNode is a wrapper around an inode, and an a inode doesn’t know where it is in the file system hierarchy. If I can get access to the Pose from the model, I can send a BMessenger to the generation thread so that it can send the message back to the pose without involving the application thread.

Patch-set: 16
Reviewer: Gerrit User 1000013 <1000013@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>