Update patch set 16
Patch Set 16:

> Patch Set 12:
> 
> > Patch Set 12:
> > I believe that my send_data() and receive_data() are not good enough/will block because the message queue is only 1 level deep. I believe that I need to create a port instead. To test this far, I need basic mime types to be present.  So far I haven’t been able to. Does anybody know where the mime type gets set and why it is saying that files are directories on foreign file systems?
> 
> A big problem I believe, is that the MIME system always sniffs a file before it tries to read its extension (which, at the moment, it doesn’t even know how to interpret extensions...).
> 
> As seen here on this ticket: https://dev.haiku-os.org/ticket/10067
> 
> There have been multiple problems with this behavior, and I believe file extensions should be checked before a file sniff is done (as on BeOS). This may fix your problem with MIME typing files from other filesystems.
> 
> I was looking into this, though I’m running low on time...the code for this is in the Registrar (check the MIMEManager class, and the classes in the “mime” folder), and private classes in the Storage Kit (check the “mime” folder).

I figured out that the all the file type detection depends on a BEOS:TYPE attribute to be there. If it’s not there, the best that it can do is figure out if it’s a file or a directory from the stat information. So if you can’t write attributes to a volume because it is read-only or doesn’t support attributes, then there is hope for thumbnails to ever work there.

However, if the BEOS:TYPE attribute is present, even if the volume is read-only, I can generate a thumbnail, and that’s good enough for me.

Falling back to looking at the extension to figure out the type of the file when the BEOS:TYPE parameter is missing would be a nice addition, but that’s out of the scope of this patch set, and what BeOS and Haiku have ever done.

Patch-set: 16
Reviewer: Gerrit User 1000013 <1000013@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>