Update patch set 16
Patch Set 16:

> Patch Set 8:
> 
> > Patch Set 8: Code-Review-1
> 
> I’ll update this to store an 8-byte creation time attribute on both platforms. However this means that 32-bit Haiku thumbnail regeneration will no longer work right after 2038. I could reproduce the overflow on 32-bit Haiku so comparisons will be in the same epoch there... but that will mean that when you copy the image files to your 64-bit Haiku after 2038 they will have stale thumbnails.

I have an idea on how to fix this. Sometime between now and 2038 we calculate an epoch in the modification time struct. We use the epoch to calculate the correct modification time for your machine. Until 2038 epoch will always be 0 on both 32-bit and 64-bit and so Haiku will behave identically to how it does now. After 2038 epoch will be calculated to be 1 on 32-bit and still 0 on 64-bit. When you ReadAttr or WriteAttr depending on the size of time_t (that you pass in) to get or set the modification attribute we can calculate the epoch and use it to and patch up the modification time that is read off disk or written to disk so that it is in the appropriate epoch for your machine. So files generated on 32-bit Haiku that have overflowed will be converted when read on 64-bit Haiku to fit in 8 bytes without overflow. Files with modification times generated on 64-bit Haiku would be converted to overflow at INT32_MAX (the biggest value of a 4 byte time_t) so that the value written would the same as a 32-bit machine would write to fit it in 4 bytes. Both platforms would continue to work and could share file’s with each other without having to regenerate thumbnails (unless the image contents change of course!)

Patch-set: 16
Reviewer: Gerrit User 1000013 <1000013@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>