Update patch set 1
Patch Set 1: -Code-Review

> Patch Set 1:
> 
> I don't think you should trust the comments in the linker script. As this linker script was from very early efi work, other OS'es weren't that good at linking EFI apps either. So I think linker scripts may have already been fixed to handle bss and still boot.
> 
> I kind of wanted to revisit it when Haiku did full EFI boots, but by then we were using GNU-EFI. So this is probably an option, but I hope we don't need it.

Well, to make things more clear here:

Currently the BSS section is stored in the executable file. This means a large part of our executable is just filled with 0. That's not how it should work. Normally we simply need to declare the size of the BSS section and the loader will clear it for us. This part of the change should be merged.

What needs to be tested is if the code in start.cpp to clear that memory ourselves is needed. It should not be the case, but I'm not sure. One possible problem is that the bss could be "lost" or incorrectly converted when generating the Pe executable, and then it would not be properly cleared. It would be nice to confirm if that's the case, and, of course, if the BSS is not cleared by the loader, try to fix that.

Patch-set: 1
Reviewer: Gerrit User 1000001 <1000001@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>
Label: Code-Review=0